Structure project refactor
