--- AP_app/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ApAppConfig(AppConfig):
-    name = 'AP_app'
