--- AP_app/tests.py
@@ -1,110 +1,0 @@
-from datetime import datetime
-from http import HTTPStatus
-from parameterized import parameterized
-from pytz import UTC
-from unittest.mock import MagicMock
-from os import path, remove
-
-from django.test import Client, TestCase
-from django.urls import reverse
-from django.core.files import File
-
-from users_app.models import User
-from invoices_app.models import (
-    Company,
-    InvoiceArg,
-    TaxPayer,
-)
-
-
-class TestAP(TestCase):
-    def setUp(self):
-        self.company = Company.objects.create(name='Company testing')
-        self.user = User.objects.create_user(email='ap@eventbrite.com')
-        self.taxpayer = TaxPayer.objects.create(
-            name='Eventbrite',
-            workday_id='12345',
-            company=self.company,
-        )
-        self.file_mock = MagicMock(spec=File)
-        self.file_mock.name = 'test.pdf'
-        self.file_mock.size = 50
-        self.invoice_creation_valid_data = {
-            'invoice_date': datetime(2007, 12, 5, 0, 0, 0, 0, UTC),
-            'invoice_type': 'A',
-            'invoice_number': '1234',
-            'po_number': '98876',
-            'currency': 'ARS',
-            'net_amount': '4000',
-            'vat': '1200',
-            'total_amount': '5200',
-            'taxpayer': self.taxpayer,
-            'user': self.user,
-            'invoice_file':self.file_mock
-        }
-        self.invoice_creation_empty_data = {}
-        self.client = Client()
-    
-    def tearDown(self):
-        if path.exists(self.file_mock.name):
-            remove(self.file_mock.name)
-
-    @parameterized.expand([
-        ('ap-home',),
-        ('ap-welcome',),
-    ])
-    def test_ap_site_permission(self, page_name):
-        self.client.force_login(self.user)
-        response = self.client.get(
-            reverse(page_name),
-            follow=True,
-        )
-        self.assertEqual(response.status_code, HTTPStatus.OK)
-
-    @parameterized.expand([
-        ('ap-home',),
-        ('ap-welcome',),
-    ])
-    def test_ap_site_access_denied(self, page_name):
-        user = User.objects.create_user(email='not_allowed@user.com')
-        self.client.force_login(user)
-        response = self.client.get(
-            reverse(page_name),
-            follow=True,
-        )
-        self.assertIn(
-            ('/suppliersite/home?next={}'.format(reverse(page_name)), 302),
-            response.redirect_chain
-            )
-        self.assertEqual(response.status_code, HTTPStatus.OK)
-        self.assertEqual(response.template_name, ['supplier_app/supplier-home.html'])
-
-    def test_ap_invoices_list_view(self):
-        self.client.force_login(self.user)
-        invoice = InvoiceArg.objects.create(**self.invoice_creation_valid_data)
-        response = self.client.get(
-            reverse('ap-home')
-        )
-        self.assertContains(response, invoice.po_number)
-        self.assertContains(response, invoice.taxpayer.name)
-
-    def test_ap_invoices_list_are_in_new_status(self):
-        self.client.force_login(self.user)
-        invoice1 = InvoiceArg.objects.create(**self.invoice_creation_valid_data)
-        other_tax_payer = TaxPayer.objects.create(
-            name='Test Tax Payer',
-            workday_id='12345',
-            company=self.company,
-        )
-        other_invoice_data = self.invoice_creation_valid_data
-        other_invoice_data['taxpayer'] = other_tax_payer
-        invoice2 = InvoiceArg.objects.create(**other_invoice_data)
-        invoice2.status = 'APPROVED'
-        invoice2.save()
-
-        response = self.client.get(
-            reverse('ap-home')
-        )
-        # Only the invoice with NEW status should be listed.
-        self.assertContains(response, invoice1.taxpayer.name)
-        self.assertNotContains(response, invoice2.taxpayer.name)
