+++ invoices_app/__init__.py
@@ -1,0 +1,17 @@
+CURRENCIES = [
+    ('ARS', 'ARS'),
+    ('USD', 'USD')
+    ]
+
+INVOICE_STATUS = [
+    ('1', 'NEW'),
+    ('2', 'APPROVED'),
+    ('3', 'CHANGES REQUEST'),
+    ('4', 'REJECTED'),
+    ('5', 'PAID'),
+]
+
+ARS_INVOICE_TYPES = [
+    ('A', 'A'),
+    ('C', 'C'),
+]
