+++ invoices_app/views.py
@@ -1,5 +1,63 @@
+from django.views.generic.list import ListView
+from django.views.generic import CreateView
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.urls import reverse_lazy
+from django.shortcuts import get_object_or_404
+from pure_pagination.mixins import PaginationMixin
+
+from invoices_app.forms import (
+    InvoiceForm
+)
+from invoices_app.models import Invoice
+from supplier_app.models import TaxPayer
 
 
+class InvoiceListView(
+    LoginRequiredMixin,
+    PaginationMixin,
+    ListView
+):
+    template_name = 'AP_app/invoice-list.html'
+    model = Invoice
+    paginate_by = 10
+
+    def get_queryset(self):
+        queryset = Invoice.objects.filter(status='NEW').order_by('id')
+        return queryset
 
 
+class SupplierInvoiceListView(LoginRequiredMixin, PaginationMixin, ListView):
+    template_name = 'supplier_app/invoice-list.html'
+    model = Invoice
+    paginate_by = 10
 
+    def get_queryset(self):
+        tax_payer = get_object_or_404(TaxPayer, id=self.kwargs['taxpayer_id'])
+        queryset = Invoice.objects.filter(taxpayer=tax_payer.id).order_by('id')
+        return queryset
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['taxpayer_id'] = self.kwargs['taxpayer_id']
+        return context
+
+
+class SupplierInvoiceCreateView(CreateView):
+    model = Invoice
+    form_class = InvoiceForm
+    template_name = 'supplier_app/invoices_form.html'
+
+    def get_success_url(self):
+        return reverse_lazy('supplier-invoice-list', kwargs={'taxpayer_id': self.kwargs['taxpayer_id']})
+
+    def form_valid(self, form):
+        form.instance.user = self.request.user
+        tax_payer = get_object_or_404(TaxPayer, id=self.kwargs['taxpayer_id'])
+        form.instance.taxpayer = tax_payer
+        self.object = form.save()
+        return super().form_valid(form)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['taxpayer_id'] = self.kwargs['taxpayer_id']
+        return context
