+++ reqdev.txt
@@ -27,6 +27,7 @@ flake8==3.7.8
 formatstring==1.1
 gunicorn==19.9.0
 idna==2.8
+ipdb==0.12.2
 ipython==7.8.0
 ipython-genutils==0.2.0
 isodate==0.6.0
