+++ started
@@ -1,0 +1,5 @@
+Script started on Mon Sep 30 12:16:20 2019
+command: on Mon Sep 30 12:15:43 2019
+Script: on: No such file or directory
+
+Script done on Mon Sep 30 12:16:20 2019
