+++ supplier_app/__init__.py
@@ -1,2 +1,7 @@
-MAX_SIZE_FILE=5242880 # 5 * 1014 * 1024
-ALLOWED_FILE_EXTENSIONS=['.pdf']
+# 5 * 1014 * 1024
+MAX_SIZE_FILE = 5242880
+ALLOWED_FILE_EXTENSIONS = ['.pdf']
+TAXPAYER_STATUS = [
+    ("PENDING", "Pending"),
+    ("ACTIVE", "Active"),
+]
