+++ templates/supplier_app/navbar.html
@@ -32,6 +32,16 @@
                 </div>
             </li>
             {% endif %}
+            <li class="nav-item dropdown">
+                <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown"
+                    aria-haspopup="true" aria-expanded="false">
+                    Management
+                </a>
+                <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+                    <a class="dropdown-item" href="{% url 'invoices-list' %}"><i
+                            class="fa fa-file-invoice mr-2"></i>{% trans "Invoices" %}</a>
+                </div>
+            </li>
 
         </ul>
         <div class="collapse navbar-collapse justify-content-end" >
