+++ users_app/__init__.py
@@ -1,0 +1,10 @@
+ALLOWED_AP_ACCOUNTS = [
+    'ap@eventbrite.com',
+    'nicolas.nunez@eventbrite.com',
+    'juan.vallejo@eventbrite.com',
+    'julian.rofrano@eventbrite.com',
+    'julian.andres@eventbrite.com',
+    'nahuel.valencia@eventbrite.com',
+    'kgerez@eventbrite.com',
+    'analaura@eventbrite.com',
+]
