+++ users_app/tests.py
@@ -1,14 +1,15 @@
+from http import HTTPStatus
+from parameterized import parameterized
+
 from django.test import TestCase, Client
 from django.contrib.auth import get_user_model
 from django.urls import reverse
-
 from social_django.models import UserSocialAuth
-from http import HTTPStatus
 
 from users_app.models import User
 
-SUPPLIER_HOME = '/suppliersite/home'
-AP_HOME = '/apsite/home'
+SUPPLIER_HOME = '/suppliersite/supplier'
+AP_HOME = reverse('ap-taxpayers')
 GENERIC_PASSWORD = '1234'
 
 
@@ -106,7 +107,7 @@ class TestLoginRedirect(TestCase):
         url_ap, status_code_ap = response_ap.redirect_chain[0]
         url_sup, status_code_sup = response_supplier.redirect_chain[0]
         self.assertEqual(HTTPStatus.FOUND, status_code_ap)
-        self.assertEqual('/suppliersite/home?next=/apsite/home'.format(AP_HOME), url_ap)
+        self.assertEqual(reverse('supplier-home')+'?next={}'.format(AP_HOME), url_ap)
         self.assertEqual(HTTPStatus.FOUND, status_code_sup)
         self.assertEqual('/?next={}'.format(SUPPLIER_HOME), url_sup)
 
@@ -114,4 +115,34 @@ class TestLoginRedirect(TestCase):
         self.client.force_login(self.user_with_google_social)
         response = self.client.get(AP_HOME, follow=True)
         self.assertEqual(HTTPStatus.OK, response.status_code)
-        self.assertEqual('AP_app/ap-home.html', response.template_name[0])
+        self.assertEqual('AP_app/ap-taxpayers.html', response.template_name[0])
+
+    @parameterized.expand([
+        ('invoices-list',),
+        ('ap-taxpayers',),
+    ])
+    def test_ap_site_permission(self, page_name):
+        self.client.force_login(self.user_with_google_social)
+        response = self.client.get(
+            reverse(page_name),
+            follow=True,
+        )
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+
+    @parameterized.expand([
+        # ('invoices-list',),
+        ('ap-taxpayers',),
+    ])
+    def test_ap_site_access_denied(self, page_name):
+        user = User.objects.create_user(email='not_allowed@user.com')
+        self.client.force_login(user)
+        response = self.client.get(
+            reverse(page_name),
+            follow=True,
+        )
+        self.assertIn(
+            ('/suppliersite/supplier?next={}'.format(reverse(page_name)), 302),
+            response.redirect_chain
+            )
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        self.assertEqual(response.template_name, ['supplier_app/supplier-home.html'])
